Add flip and spin code

+++ main.py
@@ -2,6 +2,7 @@ import lib.libardrone
 import boto3
 import boto
 from boto.s3.key import Key
+from time import sleep
 
 def searchTarget(drone):
 	c = boto.connect_s3()
@@ -21,28 +22,40 @@ def searchTarget(drone):
 		# need to turn left
 		print "2"
 	
-	if(x )
-
 def main():
 	sqs = boto3.resource('sqs', region_name="us-east-1")
 	q = sqs.get_queue_by_name(QueueName="DroneQueue")
-	q.purge()
+	try:
+		q.purge()
+	except:
+		pass
 		
 
 	drone = lib.libardrone.ARDrone(True)
 	print "Connected to Drone"
+	drone.reset()
 	done = False
 	try:
 		while not done:
+			print ".",
 			rs = q.receive_messages()
+			if not rs:
+				continue
 			for m in rs:
 				body = m.body
 				if body == "reset":
 					body = "emergency"
+				print "Got command", body
 				if body == "search":
 					searchTarget(drone)
-				drone.apply_command(body)
+				elif body == "spin":
+					drone.event_turnarround()
+				elif body == "flip":
+					drone.event_flip()
+				else:
+					drone.apply_command(body)
 				m.delete()
+			sleep(0.5)
 	except (KeyboardInterrupt, SystemExit):
 		drone.halt()
